VM_ArgPtr, VM_ArgArray: Drop if size or count argument is < 0
Before this was supposed to be checked by syscall, but TRAP_MEMCPY and
TRAP_MEMSET didn't do this.

+++ src/qcommon/qcommon.h
@@ -342,8 +342,8 @@ intptr_t QDECL VM_Call(vm_t *vm, int callnum, ...);
 
 void	VM_Debug( int level );
 
-void	*VM_ArgPtr( intptr_t intValue, intptr_t size );
-void	*VM_ArgArray( intptr_t intValue, intptr_t size, intptr_t num );
+void	*VM_ArgPtr( intptr_t intValue, uint32_t size );
+void	*VM_ArgArray( intptr_t intValue, uint32_t size, uint32_t num );
 char	*VM_ArgString( intptr_t intValue );
 intptr_t	VM_strncpy( intptr_t dest, intptr_t src, intptr_t size );
 void	VM_LocateGameDataCheck( const void *data, int entitySize, int num_entities );
